# Conflicts:
#	build.gradle
#	sitewhere-microservice-api/src/main/java/com/sitewhere/microservice/api/user/UserCreateRequest.java
#	sitewhere-microservice/src/main/java/com/sitewhere/microservice/security/SystemUser.java
#	sitewhere-microservice/src/main/java/com/sitewhere/microservice/security/TokenManagement.java

+++ sitewhere-microservice-api/src/main/java/com/sitewhere/microservice/api/device/ChartBuilder.java
@@ -43,7 +43,7 @@ public class ChartBuilder {
 
 	// Add all measurements.
 	for (IDeviceMeasurement mx : matches) {
-	    addSeriesEntry(mx.getName(), mx.getValue(), mx.getEventDate());
+	    addSeriesEntry(mx.getName(), mx.getValue().doubleValue(), mx.getEventDate());
 	}
 	// Sort entries by date.
 	List<IChartSeries<Double>> results = new ArrayList<IChartSeries<Double>>();
